+++ include/bitcal/bit_view.hpp
@@ -11,8 +11,7 @@ class const_bit_view {
 public:
     constexpr const_bit_view() noexcept = default;
 
-    constexpr const_bit_view(const std::uint64_t* data, std::size_t words) noexcept
-        : data_(data), word_count_(words) {}
+    constexpr const_bit_view(const std::uint64_t* data, std::size_t words) noexcept : data_(data), word_count_(words) {}
 
     [[nodiscard]] constexpr const std::uint64_t* data() const noexcept { return data_; }
     [[nodiscard]] constexpr std::size_t word_count() const noexcept { return word_count_; }
@@ -36,8 +35,7 @@ class bit_view {
 public:
     constexpr bit_view() noexcept = default;
 
-    constexpr bit_view(std::uint64_t* data, std::size_t words) noexcept
-        : data_(data), word_count_(words) {}
+    constexpr bit_view(std::uint64_t* data, std::size_t words) noexcept : data_(data), word_count_(words) {}
 
     [[nodiscard]] constexpr std::uint64_t* data() noexcept { return data_; }
     [[nodiscard]] constexpr const std::uint64_t* data() const noexcept { return data_; }
@@ -48,9 +46,7 @@ public:
         return data_[index];
     }
 
-    [[nodiscard]] constexpr operator const_bit_view() const noexcept {
-        return const_bit_view(data_, word_count_);
-    }
+    [[nodiscard]] constexpr operator const_bit_view() const noexcept { return const_bit_view(data_, word_count_); }
 
 private:
     std::uint64_t* data_ = nullptr;