+++ include/bitcal/bitcal.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "config.hpp"
-#include "bit_view.hpp"
-#include "bit_block.hpp"
+
 #include "algorithms.hpp"
+#include "bit_block.hpp"
+#include "bit_view.hpp"